Move the password reading classes to the shared folder, since they are not specific to the configuration.
