+++ cli/cli-api/src/test/java/com/quorum/tessera/cli/keypassresolver/CliKeyPasswordResolverTest.java
@@ -5,8 +5,8 @@ import com.quorum.tessera.config.keypairs.ConfigKeyPair;
 import com.quorum.tessera.config.keypairs.DirectKeyPair;
 import com.quorum.tessera.config.keypairs.FilesystemKeyPair;
 import com.quorum.tessera.config.keypairs.InlineKeypair;
-import com.quorum.tessera.config.util.PasswordReader;
-import com.quorum.tessera.config.util.PasswordReaderFactory;
+import com.quorum.tessera.passwords.PasswordReader;
+import com.quorum.tessera.passwords.PasswordReaderFactory;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
