+++ cli/config-cli/src/main/java/com/quorum/tessera/config/cli/parsers/KeyUpdateParser.java
@@ -7,9 +7,9 @@ import com.quorum.tessera.config.PrivateKeyData;
 import com.quorum.tessera.config.PrivateKeyType;
 import com.quorum.tessera.config.keys.KeyEncryptor;
 import com.quorum.tessera.config.util.JaxbUtil;
-import com.quorum.tessera.config.util.PasswordReader;
 import com.quorum.tessera.encryption.PrivateKey;
 import com.quorum.tessera.io.SystemAdapter;
+import com.quorum.tessera.passwords.PasswordReader;
 import org.apache.commons.cli.CommandLine;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
