+++ cli/config-cli/src/test/java/com/quorum/tessera/config/cli/parsers/KeyUpdateParserTest.java
@@ -3,8 +3,8 @@ package com.quorum.tessera.config.cli.parsers;
 import com.quorum.tessera.config.*;
 import com.quorum.tessera.config.keys.KeyEncryptorFactory;
 import com.quorum.tessera.config.util.JaxbUtil;
-import com.quorum.tessera.config.util.PasswordReader;
 import com.quorum.tessera.encryption.PrivateKey;
+import com.quorum.tessera.passwords.PasswordReader;
 import org.apache.commons.cli.*;
 import org.junit.Before;
 import org.junit.Test;
