+++ key-generation/src/test/java/com/quorum/tessera/key/generation/FileKeyGeneratorTest.java
@@ -6,11 +6,11 @@ import com.quorum.tessera.config.PrivateKeyData;
 import com.quorum.tessera.config.PrivateKeyType;
 import com.quorum.tessera.config.keypairs.FilesystemKeyPair;
 import com.quorum.tessera.config.keys.KeyEncryptor;
-import com.quorum.tessera.config.util.PasswordReader;
 import com.quorum.tessera.encryption.KeyPair;
 import com.quorum.tessera.encryption.PrivateKey;
 import com.quorum.tessera.encryption.PublicKey;
 import com.quorum.tessera.nacl.NaclFacade;
+import com.quorum.tessera.passwords.PasswordReader;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
