+++ shared/src/main/java/com/quorum/tessera/passwords/PasswordReaderFactory.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.config.util;
+package com.quorum.tessera.passwords;
 
 import java.io.Console;
